+++ packages/browser-wallet/src/popup/pages/Account/TransactionLog/i18n/da.ts
@@ -9,6 +9,7 @@ const t: typeof en = {
     events: 'Hændelser',
     rejectReason: 'Afvisningsårsag',
     error: 'Fejl ved hentning af transaktionshistorik',
+    noTransactions: 'Denne konto har ingen transakationer endnu.',
 };
 
 export default t;